Provide a generic build script for the sigrok stack
build-sigrok.sh is a shell script for building the sigrok stack
either natively or in cross-compile mode. It does not clone any
repositories by itself, but builds out-of-source from existing
local clones. This makes it useful for hacking.

The details of the build can be controlled via command-line
options. The default configuration is sourced from a environment
script for the selected host platform.